requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.3.6
gevent==24.2.1
gevent-websocket==0.10.1
python-socketio==5.11.1
python-engineio==4.9.0
mysql-connector-python==8.3.0
Werkzeug==3.0.1
PyJWT==2.8.0
bcrypt==4.1.2
